requirements.txt
colour
matplotlib
numpy
pandas
pillow
pytest
requests
scikit-image
scipy
torch
torchvision
tqdm
yt-dlp
pyyaml
rich
lightning
pydantic
graphkit
einops
imageio
opencv-python
pydantic
PyYAML
scikit-learn
segmentation_models_pytorch
timm
rich